+++ src/main/java/org/violetmoon/quark/content/tweaks/module/CampfiresBoostElytraModule.java
@@ -1,6 +1,7 @@
 package org.violetmoon.quark.content.tweaks.module;
 
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
@@ -12,6 +13,7 @@ import net.minecraft.world.level.block.CampfireBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.Vec3;
 
+import org.violetmoon.quark.base.Quark;
 import org.violetmoon.quark.content.building.module.GrateModule;
 import org.violetmoon.zeta.config.Config;
 import org.violetmoon.zeta.event.bus.PlayEvent;